@@ -1,16 +1,15 @@
 {
   "author": "tony_quetano@planttheidea.com",
   "ava": {
-    "babel": "inherit",
     "failFast": true,
     "files": [
       "test/*.js"
     ],
     "require": [
-      "babel-register",
+      "@babel/register",
       "test/helpers/setup-browser-env.js"
     ],
-    "source": [
+    "sources": [
       "src/*.js"
     ],
     "verbose": true
@@ -28,37 +27,39 @@
   },
   "description": "Handle react classes with more functional purity",
   "devDependencies": {
-    "ava": "^0.25.0",
-    "babel-cli": "^6.26.0",
-    "babel-eslint": "^8.2.6",
-    "babel-loader": "^7.1.5",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "babel-register": "^6.26.0",
+    "@babel/cli": "^7.0.0",
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.0.0",
+    "ava": "^1.0.0-beta.8",
+    "babel-eslint": "^9.0.0",
+    "babel-loader": "^8.0.0",
     "browser-env": "^3.2.5",
-    "eslint": "^5.2.0",
-    "eslint-config-rapid7": "^3.0.3",
+    "eslint": "^5.5.0",
+    "eslint-config-rapid7": "^3.1.0",
     "eslint-friendly-formatter": "^4.0.1",
     "eslint-loader": "^2.1.0",
     "html-webpack-plugin": "^3.0.7",
     "in-publish": "^2.0.0",
-    "nyc": "^12.0.2",
+    "micro-memoize": "^2.1.0",
+    "nyc": "^13.0.1",
     "optimize-js-plugin": "^0.0.4",
     "prop-types": "^15.6.2",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
-    "react-hot-loader": "^4.3.4",
-    "rollup": "^0.63.4",
-    "rollup-plugin-babel": "^3.0.7",
-    "rollup-plugin-commonjs": "^9.1.4",
-    "rollup-plugin-node-resolve": "^3.3.0",
-    "rollup-plugin-uglify": "^4.0.0",
-    "sinon": "^6.1.4",
-    "styled-components": "^3.3.3",
-    "webpack": "^4.16.3",
+    "react": "^16.5.0",
+    "react-dom": "^16.5.0",
+    "react-hot-loader": "^4.3.6",
+    "rollup": "^0.65.2",
+    "rollup-plugin-babel": "^4.0.1",
+    "rollup-plugin-commonjs": "^9.1.6",
+    "rollup-plugin-node-resolve": "^3.4.0",
+    "rollup-plugin-uglify": "^5.0.2",
+    "sinon": "^6.3.1",
+    "styled-components": "^3.4.6",
+    "webpack": "^4.18.0",
     "webpack-cli": "^3.1.0",
-    "webpack-dev-server": "^3.1.5"
+    "webpack-dev-server": "^3.1.8"
   },
   "keywords": [
     "functional",
