@@ -41,7 +41,7 @@ module.exports = Object.assign({}, defaultConfig, {
       if (rule.loader === 'babel-loader') {
         return Object.assign({}, rule, {
           options: Object.assign({}, rule.options, {
-            presets: ['react'],
+            presets: ['@babel/react'],
           }),
         });
       }
